initial commit. file structure
prompt to copilot: make a sticky notes app that is web based. i am going to use react and the app should be a website. can you create an outline for me. i dont want code i just want the files and maybe neceseary code to link the files. kept code we understood